+++ application/modules/events/controllers/admin/Index.php
@@ -1,4 +1,5 @@
 <?php
+
 /**
  * @copyright Ilch 2
  * @package ilch
@@ -45,8 +46,7 @@ class Index extends \Ilch\Controller\Admin
             $items[0]['active'] = true;
         }
 
-        $this->getLayout()->addMenu
-        (
+        $this->getLayout()->addMenu(
             'menuEvents',
             $items
         );
@@ -65,12 +65,12 @@ class Index extends \Ilch\Controller\Admin
             }
         }
 
-        $this->getView()->set('event', $eventMapper->getEntries());
+        $this->getView()->set('events', $eventMapper->getEntries());
     }
 
     public function delAction()
     {
-        if ($this->getRequest()->isSecure()) {
+        if ($this->getRequest()->isSecure() && $this->getRequest()->getParam('id')) {
             $eventMapper = new EventMapper();
             $eventMapper->delete($this->getRequest()->getParam('id'));
